#295: Use upstream version of the monomorphizer  (#495)
Instead of keeping our own copy of the monomorphizer, we would like to
start using rustc monomorphizer as is. This is the first step towards
that goal.

For now, we will use the upstream version of the monomorphizer with a
few small changes. Use a query to skip monomorphization for rmc instead
of hooks.

This change, adds a query to allow RMC to skip monomorphizing functions. It also does the following:
- Adds a default query to codegen base that don't skip anything.
- Adds a query implementation to RMC that checks our hooks.
- Skips monomorphizing the start function if --cfg=rmc

Issue #485 tracks the rest of the changes needed.