Revert some changes to rust compiler code (#42) (#667)
* Use rustc SwitchTarget as is (#42)

There is no need to modify rustc code in order to access the
SwitchTarget targets field. Change direct access to calls to
all_targets().

* Remove old change made to rustc_driver (#42)

This is no longer needed.

* Use rustc TyS as is (#42)

There is no need to access the kind and flags field of TyS directly.
We should be using kind() and flags() methods instead. Thus, I'm
removing the modification we made to the original rustc code to make
these variables public and fixing the code that relied on the direct
access.

* Revert changes from ScalarInt

Use conversion methods instead.