Split rmc::nondet into safe and unsafe functions(#607) (#657)
* Introduce rmc::any and rmc::any_raw and deprecate rmc::nondet

We decided to replace the old method rmc::nondet because its behavior
wasn't clear. For some types it would respect the type invariants but
not for others. Additionally, there was no way for the user to specify
how to create safe non-deterministic values.

Users will be able to control how RMC generate symbolic values to their
variables.

- `rmc::any::<T>()` will provide a safe way to generate values
that respect any invariant specified to type T (Given by the
implementation of Arbitrary trait.)

- `rmc::any_raw::<T>()` will provide any possible bit value for the
memory layout of the type. This is an unsafe method and users must
ensure correctness after this call.

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>
Co-authored-by: Adrian Palacios <73246657+adpaco-aws@users.noreply.github.com>